requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scikit-image==0.22.0
